Rename all pages to title case
