+++ src/large/SegmentMain.js
@@ -1,5 +1,5 @@
 import React from "react"
-import { Row, Col } from "reactstrap"
+import { Col } from "reactstrap"
 
 const SegmentMain = props => <Col id="SegmentMain">{props.children}</Col>
