+++ src/medium/ActionButtons.js
@@ -4,7 +4,7 @@ import ActionButton from "../small/ActionButton"
 
 const ActionButtons = props => (
   <span className="action buttons">
-    <ActionButton to="/create">Create Post</ActionButton>
+    <ActionButton to="/new">Write New Post</ActionButton>
     <ActionButton to="/register">Register</ActionButton>
     <ActionButton to="/login">Login</ActionButton>
   </span>
