+++ src/pages/Register.js
